feat: elias-memory v0.2.0 — Desktop/Mobile 2-tier architecture
New features:
- Desktop/Mobile profiles (auto-detect based on available libs)
- Memory consolidation (episodic → semantic, topic-based + vector clustering)
- Knowledge gap detection (coverage analysis + retrieval gap check)
- Memory deletion and stats API
- Fixed HashEmbedder overflow (uint8 scaling instead of raw float32 reinterpret)
- Fixed numpy matmul overflow (cast to float32)
- Decay cycle now returns pruned count and auto-deletes dead memories
- SFT export with min_importance filter

Architecture:
- Desktop: NVIDIA embeddings + FAISS index (when available)
- Mobile: Hash embeddings + numpy brute-force (always works)
- Same API, different backends via profile parameter

44 tests passing.

+++ src/elias_memory/store/vec.py
@@ -28,7 +28,7 @@ class NumpyVectorIndex(VectorIndex):
         norm = np.linalg.norm(query_vec)
         q = query_vec / norm if norm > 0 else query_vec
         matrix = np.stack(self._vecs)
-        similarities = matrix @ q
+        similarities = matrix.astype(np.float32) @ q.astype(np.float32)
         top_indices = np.argsort(similarities)[::-1][:top_k]
         return [(self._ids[i], float(similarities[i])) for i in top_indices]
